requirements.txt
# Backend API Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiohttp==3.10.10

# Environment & Configuration
python-dotenv==1.0.1

# Database & Storage
supabase==2.9.0

# LiveKit for voice agents
livekit==0.17.4
livekit-api==0.6.2
livekit-agents[openai,turn-detector,silero,cartesia,deepgram]==1.2.7
livekit-plugins-noise-cancellation==0.2.2

# OpenAI Integration
openai>=1.0.0
tiktoken>=0.5.0

# Additional tools
requests==2.32.3
